Fix cloud-images.rst generation script
The script failed at some point back in April and generated an empty
list of cloud images that was not noticed until now. This patch fixes
that script so that we have documented supported images again.

The macro lf-infra-pre-build is able to install all the tools we need
for this job so utilize that rather than maintaining our own virtualenv
and openstack installation.

Issue: RELENG-153
Change-Id: I49c750e8051d605bf8526c73b6c8bd808f05a550
Signed-off-by: Thanh Ha <zxiiro@gmail.com>