Update patch set 6
Patch Set 6:

Noticed another condition we need to test. Some jobs in the integration project have a csit-list parameter that pulls in a file containing a list of csit jobs.

We need to duplicate that file and update the parameter appropriately as a step after the job streams are updated.

Patch-set: 6
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>