Install Docker in regular builder
Considering that many projects are utilizing Docker now and the only
real difference between the "docker" template and the "builder"
template is whether or not they have docker installed. Let's just
consolidate the 2 image types into a single "builder" one and reduce
the number of special images we are managing.

This also configures the vm flavor packer uses to build to v2-highcpu-4
to make our packer builds faster. We used limit to the lowest size
to ensure that the smallest volume is used however considering we are
explicitly passing a vm_volume_size parameter now let's not limit
ourselves to the CPU size of our build system.

Change-Id: I116ab774b27ddedd5bf55746eb2cf067256797e0
Signed-off-by: Thanh Ha <zxiiro@gmail.com>