Feat: Add Job template to deploy ODL on K8S Magnum
The template allows users to create a K8S cluster on Openstack Magnum
and deploy ODL SDNC on the cluster. Once deployed the Jenkins node
verifies the deployment by listing the restconf modules.

The JJB templates uses pre-existing K8s cluster templates provided
by the cloud provider. (v1.16.0)

The job templates are created in generic manner such that users can
pass a script to test the SDNC setup from the project definition.

The cluster is configured on a private network using internal IP's
and the Jenkins executor node IP is added to the network to allow
kubectl commands to be run directly the Jenkins instance.

Issue-ID: https://jira.linuxfoundation.org/browse/RELENG-3644
Change-Id: Ie17340da5677bb26b0cb0f673b90018b7033fa60
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>