Feat: Add template to deploy ODL on K8S magnum
The template allows create of a K8S cluster using Openstack Magnum
client and deploys ODL SDNC on the cluster. Once deployed the
Jenkins node verifies the deployment by listing the restconf
modules.

The cluster creation is based on the pre-existing templates provided
by the cloud provider.

The templates are created in generic way such that users can
passing a script to test the SDNC setup from the
project definition.

The cluster is setup on a private network using internal IP
and the Jenkins minion is added to the network to execute
Kubectl commands directly from the Jenkins instance.

Issue-ID: LF-Jira: RELENG-3644
Change-Id: Ie17340da5677bb26b0cb0f673b90018b7033fa60
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>