Feat: Add template deploy ODL on K8S magnum
This template allow a job create a ceo cluster using Openstack Magnum
client and deploys and instance of ODL SDNC on K8S cluster.
Once deploy the Jenkins node verifies the deployment by listing
the restconf modules.

The cluster creation is based on the pre-existing templates provided
by the cloud provider.

The cluster is setup on a private network using internal IP
and the Jenkins minion is added to the network to execute
Kubectl commands directly from the Jenkins instance.

Issue-ID: LF-JIRA
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>
Change-Id: Ie17340da5677bb26b0cb0f673b90018b7033fa60