CI: Self-service job to lock/unlock GERRIT_BRANCH
Add self-service job to lock or unlock a Gerrit branch for
release work or during a code-freeze. The job also can enable or
disable supercommitter rights for any given branch.

The job can be triggered by any releng committers or
release manager with this reduces the need for opening
helpdesk tickets for branch locking.

The job can be triggered to accomplish the following:
1. Code freeze
2. Super committer rights
3. Release work
4. Unlock (removes the section for any of the above)

Issue-ID: LF-JIRA RELENG-4128
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>
Change-Id: I85d1e71c1cb972b4f51f4ac8485953fae67a92eb