Add odlparent verify/JDK17 jobs
We need to start verifying against JDK17 as we prepare to make the
switch to requiring JDK17.

Add a odl-maven-verify-jobs-jdk17 template and use it for
odlparent/master in addition to normal odl-maven-verify-jobs.

Since JDK17 is only available on CentOS8, also switch the build node
from default (which is still CentOS7).

JIRA: ODLPARENT-256
Change-Id: I4caa06d0e80c3e2a8c7e71a965af6202adca167d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>