Fix: common-functions.sh sourced w/ arg bashisms
jjb/integration/common-functions.sh was always called with source and
an argument up to this change. Contrary to the source bash syntax,
the dot POSIX equivalent syntax does not allow any argument.
The only argument taken here is the name of the bundle folder used as a
variable in the script to set a few paths.
Thus, when migrating to the dot syntax, we have to ensure the variable
${BUNDLEFOLDER} is correctly set. Hopefully, the global and local names
used at each call for these variables were identical here.

Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: Ia63c7dc33efca1f53ba8413fec762befae4a7193