Fix: sourced files w/o arg bashisms
The "source" command is only supported by a limited number of shells.
It is normally equivalent to the POSIX "." (pronounced "dot") command.
Contrary to sourced files, dotted files in bash does not allow any
extra argument since it is normally not supported in POSIX.
If no argument is specified, "source" and "." are strictly equivalent.
And sourced files can be straightforward migrated to the dot syntax.

Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: Id902674c3eef31ce9633c2fc4fc2678da428f832