Fix: Use lf-activate-venv to install openstack dep
Some of the platform/flavors does not include most reecnt python update
as system packages, so use pyenv (through lf-activate-venv) for
most recent version of python 3.x.

Use of python-tools-install.sh for the pre/post build is not recommended
approach for installing python dependencies since this installs the
dependencies with `--user` option (removed in I821a86ac3b54f284e8)
and ends up with pinning depenendencies with older packages.

Instead use lf-activate-venv to setup an venv and pull in the required
dependencies and save the path of the virtualenv in a temp file that
can be checked before attempting to create a venv.

Change-Id: I00bbf08921a3ddc6349e74e9cd9d6316b0876749
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>