Fix: Update git-review to 2.3.1
The previous version of git-review is incompatible with the latest
version of git due to renaming flags.

Error:

Errors running git rebase -p -i remotes/gerrit/master
fatal: --preserve-merges was replaced by --rebase-merges

This is fixed in 2.2.0, upgrade to 2.3.1 its more recent.

Address the submodule update issues affecting the autorelease-* jobs:

git-review tries to copy commit-msg hook to submodules with incorrect
source file path (.git/hooks/commit-msg) and fails - the path should
be ../.git/hooks/commit-msg if a relative path is used since the copy
command is run in the submodule directory

lf-activate-venv creates a virtual environment in the current working
directory where lf-activate-venv is run. This clutters the repository
and all the files for the virtual environment are added for update.

To address the bug of git-review set 'core.hooksPath' with the absolute
path of the top-level hooks directory so that the correct source path
can be used regardless of the working directory.

Ref: https://review.opendev.org/c/opendev/git-review/+/818219
Ref: https://gerrit.linuxfoundation.org/infra/c/releng/global-jjb/+/70648
Change-Id: I614ccf4b22343cd234e6a63a79b97252eae4abd6
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>