Submit release note to the same branch
The release note created by 'genenrate-release-notes.sh' is pushed to
the master branch. Make sure the change is submitted to the same branch
in Gerrit.

And remove the sed command to append a new release note file in the
'index.rst' file since the file already includes a wildcard match
'release-notes-*'.

Change-Id: Id396d8fec4eae97b627fa490ea20b2c5f437b998
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>