Fix Sonar jobs
Sonar jobs for the projects that require Java 17 are failing because
the default value of 'openjdk11' configured in the global-jjb is used
for all the Sonar jobs.

Create and use two job groups for Sonar jobs, one for openjdk11 and the
other for openjdk17, which have 'java-version' explicitly specified to
fix the Java version issue.
The job groups also capture other common parameters shared across all
the jobs to make the configuration simpler.

Change-Id: I9afe72d35304b3ddbd30d23dc797f71f2a30a5b5
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>