Feat: Enable Github2Gerrit workflow
The Github2Gerrit workflow extracts the commits from a GitHub
pull-request and submits them to an upstream Gerrit repository.
This allows GitHub developers to contribute to Gerrit-based
repositories that are primarily maintained on Gerrit servers
and replicated onto GitHub.

Now enable the workflow to test with releng/builder.

Issue: RELENG-4569
Change-Id: I4141195d7d6c10ce0516888275e3bfca8e25c1b0
Ref: https://github.com/askb/github2gerrit
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>