Fix: $_ does not expand as expected on Ubuntu
The builder node are updated to Ubuntu as default.
The .bashrc file does not process '$_' expanding
the output correctly as provided in previous line.
Therefore explicitly use the dirname.

This results in the following error:
cp: target '/usr/bin/java' is not a directory

The reason this happens is because Ubuntu "/bin/sh"
points to "/bin/dash" while CentOS / RedHat points to
"/bin/bash".

Change-Id: Ib29b0b675b247b7f20fd8b25c776c36b5856b3df
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>