Fix clustering problem with missing script
To form a cluster in integraion tests configure_cluster.sh
script is used. This script is present only in the
integration-distribution project. If the integration test job is run
against other project build like netconf it fails to form a cluster,
since it is missing. To mitigate this issue, download the missing script
from integration-distribution based on the current stream version.

JIRA: RELENG-168
Change-Id: Id9a1e2601b1abd395cabf81b1aef620420b4bf46
Signed-off-by: Martin Balaz <martin.balaz@pantheon.tech>