Download missing configure_cluster.sh script
Integration test jobs rely on the 'configure_cluster.sh'script
to form a cluster. This script is packaged only within
the integration-distribution project. When a build from other project
(e.g. netconf) is used for testing, this script is missing and it
fails to form a cluster.
To mitigate this issue, download the missing 'configure_cluster.sh'
script from the integration-distribution repository based on
the current stream version.

JIRA: RELENG-168
Change-Id: Id9a1e2601b1abd395cabf81b1aef620420b4bf46
Signed-off-by: Martin Balaz <martin.balaz@pantheon.tech>