CI: Fix gerrit-change-info action
The call-gerrit-change-info job was failing because the
lfreleng-actions/gerrit-change-info action requires a
gerrit_known_hosts input parameter that was not provided. This
parameter is needed for SSH host verification when connecting
to the Gerrit server.

Added Dependabot-compatible version comment with commit date
for better dependency tracking. Tested with 'gh act' to ensure
the workflow structure is correct and all steps complete successfully.

Change-Id: I0ff40bc7656066976e5c71667d9696002e73b88c
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>