CI: Fix gerrit-change-info action add missing para
The call-gerrit-change-info job was failing because the
lfreleng-actions/gerrit-change-info action requires a
gerrit_known_hosts input parameter that was not provided.
This parameter is needed for SSH host verification when
connecting to the Gerrit server.

Also updated the action to use the latest commit (88f5758)
instead of the older commit.

Change-Id: I7d80e47fa760be31c33f74f2a88bb176808c0807
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>