CI: Remove unused devstack packer jobs
The devstack packer jobs are not actively used by any jobs in
the system. After thorough analysis of job definitions and the
job list from April 2025, no jobs were found that reference the
devstack images.

This change:
1. Removes all devstack packer job definitions:
   - packer-devstack-jobs
   - packer-devstack-pre-pip-queens-jobs
   - packer-devstack-pre-pip-rocky-jobs
   - packer-devstack-pre-pip-stein-jobs

2. Marks the related devstack image variables in defaults.yaml as unused
   but keeps them for reference.

The Heat templates are not affected as they use generic parameters for
images that are supplied at runtime.

Change-Id: I74b300c8aac1f5765875e2f8fe1252dab999532d
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>