CI: Disable RTD jobs for docs master branch
Disable the RTD verify and merge jobs for the docs master branch
(vanadium stream) by setting disable-job: true. The RTD jobs will
continue to run on stable branches (titanium and scandium).

Tox verify jobs remain enabled on all branches including master.

This is disable to test the GHA rtdv3 workflows which are enabled
on the repo.

Change-Id: I09340a2b01a4e015214b4a24e9553f04914286d1
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>