fix: Downgrade Packer version to 1.9.1
Downgrade Packer from 1.11.2 to 1.9.1 to maintain compatibility with
existing packer templates that don't have docker and ansible plugins
defined in required_plugins block.

Packer 1.9.1 still has docker and ansible as built-in, while newer
versions require them to be explicitly declared as plugins.

This matches the version that was previously working for building
packer images.

Change-Id: I41dacf7cf139f23fe2cf3d4ad75540c6e17fe8a9
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>